Add the required files for setup.
Add a new file with just stored procedures used over development of the application. In deployment and testing by other persons is recommended to use this file instead of the generic SP.sql * VERY IMPORTANT TO BE RUN FIRST *
Add new required dml for seting up the application in a new deployemnt. * VERY IMPORTANT TO BE RUN FIRST *